Use write block on archive indices
