+++ server/src/main/java/org/elasticsearch/snapshots/RestoreService.java
@@ -1609,6 +1609,7 @@ public class RestoreService implements ClusterStateApplier {
             Settings.builder()
                 .put(snapshotIndexMetadata.getSettings())
                 .put(IndexMetadata.SETTING_INDEX_VERSION_COMPATIBILITY.getKey(), clusterState.getNodes().getSmallestNonClientNodeVersion())
+                .put(IndexMetadata.SETTING_BLOCKS_WRITE, true)
         );
         // TODO: _routing? Perhaps we don't need to obey any routing here as stuff is read-only anyway and get API will be disabled
         return convertedIndexMetadata.build();
