+++ x-pack/plugin/old-lucene-versions/src/internalClusterTest/java/org/elasticsearch/xpack/lucene/bwc/ArchiveLicenseIntegTests.java
@@ -13,14 +13,8 @@ import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotR
 import org.elasticsearch.action.admin.cluster.snapshots.restore.RestoreSnapshotResponse;
 import org.elasticsearch.cluster.ClusterState;
 import org.elasticsearch.cluster.health.ClusterHealthStatus;
-import org.elasticsearch.cluster.metadata.IndexMetadata;
 import org.elasticsearch.cluster.metadata.Metadata;
-import org.elasticsearch.cluster.metadata.RepositoryMetadata;
-import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.util.BigArrays;
-import org.elasticsearch.env.Environment;
-import org.elasticsearch.indices.recovery.RecoverySettings;
 import org.elasticsearch.license.DeleteLicenseAction;
 import org.elasticsearch.license.License;
 import org.elasticsearch.license.LicensesMetadata;
@@ -29,29 +23,12 @@ import org.elasticsearch.license.PostStartBasicRequest;
 import org.elasticsearch.license.PostStartTrialAction;
 import org.elasticsearch.license.PostStartTrialRequest;
 import org.elasticsearch.license.PostStartTrialResponse;
-import org.elasticsearch.plugins.Plugin;
-import org.elasticsearch.plugins.RepositoryPlugin;
 import org.elasticsearch.protocol.xpack.XPackUsageRequest;
 import org.elasticsearch.protocol.xpack.license.DeleteLicenseRequest;
-import org.elasticsearch.repositories.IndexId;
-import org.elasticsearch.repositories.Repository;
-import org.elasticsearch.repositories.RepositoryData;
-import org.elasticsearch.repositories.fs.FsRepository;
-import org.elasticsearch.snapshots.AbstractSnapshotIntegTestCase;
-import org.elasticsearch.snapshots.SnapshotId;
 import org.elasticsearch.snapshots.SnapshotRestoreException;
-import org.elasticsearch.snapshots.mockstore.MockRepository;
-import org.elasticsearch.test.ESIntegTestCase;
-import org.elasticsearch.xcontent.NamedXContentRegistry;
 import org.elasticsearch.xpack.core.action.XPackUsageFeatureAction;
 import org.elasticsearch.xpack.core.action.XPackUsageFeatureResponse;
 import org.elasticsearch.xpack.core.archive.ArchiveFeatureSetUsage;
-import org.junit.Before;
-
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Map;
 
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertAcked;
 import static org.hamcrest.Matchers.containsString;
@@ -59,78 +36,7 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.instanceOf;
 import static org.hamcrest.Matchers.oneOf;
 
-@ESIntegTestCase.ClusterScope(supportsDedicatedMasters = false, numClientNodes = 0, scope = ESIntegTestCase.Scope.TEST)
-public class ArchiveLicenseIntegTests extends AbstractSnapshotIntegTestCase {
-
-    @Override
-    protected Collection<Class<? extends Plugin>> nodePlugins() {
-        return Arrays.asList(LocalStateOldLuceneVersions.class, TestRepositoryPlugin.class, MockRepository.Plugin.class);
-    }
-
-    public static class TestRepositoryPlugin extends Plugin implements RepositoryPlugin {
-        public static final String FAKE_VERSIONS_TYPE = "fakeversionsrepo";
-
-        @Override
-        public Map<String, Repository.Factory> getRepositories(
-            Environment env,
-            NamedXContentRegistry namedXContentRegistry,
-            ClusterService clusterService,
-            BigArrays bigArrays,
-            RecoverySettings recoverySettings
-        ) {
-            return Map.of(
-                FAKE_VERSIONS_TYPE,
-                metadata -> new FakeVersionsRepo(metadata, env, namedXContentRegistry, clusterService, bigArrays, recoverySettings)
-            );
-        }
-
-        // fakes an old index version format to activate license checks
-        private static class FakeVersionsRepo extends FsRepository {
-            FakeVersionsRepo(
-                RepositoryMetadata metadata,
-                Environment env,
-                NamedXContentRegistry namedXContentRegistry,
-                ClusterService clusterService,
-                BigArrays bigArrays,
-                RecoverySettings recoverySettings
-            ) {
-                super(metadata, env, namedXContentRegistry, clusterService, bigArrays, recoverySettings);
-            }
-
-            @Override
-            public IndexMetadata getSnapshotIndexMetaData(RepositoryData repositoryData, SnapshotId snapshotId, IndexId index)
-                throws IOException {
-                final IndexMetadata original = super.getSnapshotIndexMetaData(repositoryData, snapshotId, index);
-                return IndexMetadata.builder(original)
-                    .settings(
-                        Settings.builder()
-                            .put(original.getSettings())
-                            .put(
-                                IndexMetadata.SETTING_INDEX_VERSION_CREATED.getKey(),
-                                metadata.settings()
-                                    .getAsVersion("version", randomBoolean() ? Version.fromString("5.0.0") : Version.fromString("6.0.0"))
-                            )
-                    )
-                    .build();
-            }
-        }
-    }
-
-    private static final String repoName = "test-repo";
-    private static final String indexName = "test-index";
-    private static final String snapshotName = "test-snapshot";
-
-    @Before
-    public void createAndRestoreArchive() throws Exception {
-        createRepository(repoName, TestRepositoryPlugin.FAKE_VERSIONS_TYPE);
-        createIndex(indexName);
-        createFullSnapshot(repoName, snapshotName);
-
-        assertAcked(client().admin().indices().prepareDelete(indexName));
-
-        PostStartTrialRequest request = new PostStartTrialRequest().setType(License.LicenseType.TRIAL.getTypeName()).acknowledge(true);
-        client().execute(PostStartTrialAction.INSTANCE, request).get();
-    }
+public class ArchiveLicenseIntegTests extends AbstractArchiveTestCase {
 
     public void testFeatureUsage() throws Exception {
         XPackUsageFeatureResponse usage = client().execute(XPackUsageFeatureAction.ARCHIVE, new XPackUsageRequest()).get();
